feat: New URL Load Balancer endpoints (#20)
Implements the new URL Load Balancing feature that receives a number of URLs and creates a short URL based on them. When accessing this short URL, it redirects the user to the first valid long URL based on a random selection.
The validation of the URLs is performed asynchronously, using Domain Events.